refactor(api service): remove websocket.Conn from BaseResponse in response.go
If programmer wants to new a service object, it is necessary to add conn to it. This way is quite tedious.

Just let every Response method of WebSocketResponse return itself, and then when giving a response to client the only thing is to write the return result to WriteJson.

So in the future, any one wants to create a new response, the response must implement the Response method and return itself.